move ast parsing out of the cli
use new settings access
